langchain-course/ice_breaker/pyproject.toml
[project]
name = "ice-breaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.12",
    "langchain-community>=0.3.23",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.2",
    "tweepy>=4.15.0",
]
